Crate veeks_millis

veeks_millis

2021 new date time units and formats: veeks and millis
repo; version: 1.0.2 date: 2021-02-03 authors: Luciano Bestia

Proposal for a new date and time units and formats

Veeks and millis are my suggestion for new date-time units and formats.
Here is a long read TL;DR about the date-time reform proposal:
https://github.com/LucianoBestia/new_date_time_units_and_formats/

A super short explanation:
Years (beginning and end) remain the same as in the CE - common era calendar.
Months are obsolete and they are not used in the new format.
Veek is similar to week. One year has 52 full veeks with 7 days.
The exception is the last 53rd veek that has only 1 celebration day. For leap years there are 2 celebration days.
The new year always starts with 01v 1d - short pronunciation one-vee one-dee. Basically every year starts with a monday.
Veek-days are similar to week-days, but without names. They use just numbers and the d unit.
1d is the new name for Monday and 7d is the new name for Sunday.
The veek_date format is global for every language and looks exactly like this:
2021c 52v 2d
The year must have 4 digits from 1000c to 9999c. The unit c stands for CE - common era. Space.
The veek must have 2 digits from 01v to 53v. The unit v stands for veek. Space.
The day must have 1 digit from 1d to 7d. The unit d stands for day. End.

Hours, minutes and seconds are obsolete.
One day is divided into 1000md.
Millis is the short name for milliday. The unit is md.
For shorter time intervals there is microdays or micros, unit μd. 1md has 1000μd.

This crate contains functions to use with the new units veeks and millis.
It is dependent on the crate chrono for NaiveDate and NaiveTime.

Used in projects

https://github.com/LucianoBestia/new_date_time_units_and_formats/

cargo crev reviews and advisory

It is recommended to always use cargo-crev
to verify the trustworthiness of each of your dependencies.
Please, spread this info.
On the web use this url to read crate reviews. Example:
https://web.crev.dev/rust-reviews/crate/num-traits/

Modules

micro_time_mod
micro_time_mod

milli_time_mod
milli_time_mod

veek_date_mod
veek_date_mod

Structs

MicroTime
micro_time format with decimals: 12.34μd, decimal digits, no space, suffix μd

MilliTime
milli_time format is fixed and global: from 000md to 999md, always 3 digits, no space, suffix md

VeekDate
veek_date format is fixed and global:
4 digit year has the unit c for CE common era, one space
2 digit for veek (very similar to week) from 01v to 53v, has unit v, one space
1 digit for veek-day from 1d (monday) to 7d(sunday)
example: 2021c 53v 7d