Struct veeks_millis::VeekDate

pub struct VeekDate {
    year: u32,
    veek: u32,
    day: u32,
}

veek_date format is fixed and global:
4 digit year has the unit c for CE common era, one space
2 digit for veek (very similar to week) from 01v to 53v, has unit v, one space
1 digit for veek-day from 1d (monday) to 7d(sunday)
example: 2021c 53v 7d

Fields
year: u32veek: u32day: u32
Implementations

impl VeekDate

pub fn from_yvd(year: u32, veek: u32, day: u32) -> Option<Self>

constructor from year, veek, day returns None on error

Example

let veeks = veeks_millis::VeekDate::from_yvd(2021,09,3).unwrap();
assert_eq!(veeks.to_string(),"2021c 09v 3d");

pub fn from_year_ordinal_day(year: u32, ordinal_day: u32) -> Option<Self>

constructor from year, ordinal_day returns None on error

Example

let veeks = veeks_millis::VeekDate::from_year_ordinal_day(2021,45).unwrap();
assert_eq!(veeks.to_string(),"2021c 07v 3d");

pub fn from_naive_date(nd: NaiveDate) -> Option<Self>

constructor from NaiveDate

Example

let nd = chrono::NaiveDate::from_ymd(2021,02,28);
let veeks = veeks_millis::VeekDate::from_naive_date(nd).unwrap();
assert_eq!(veeks.to_string() ,"2021c 09v 3d");

pub fn from_ymd(year: u32, month: u32, day: u32) -> Option<Self>

constructor from year, month, day returns None on error

Example

let veeks = veeks_millis::VeekDate::from_ymd(2021,02,28).unwrap();
assert_eq!(veeks.to_string(),"2021c 09v 3d");

pub fn from_str(s: &str) -> Option<Self>

constructor from str in VeekDate format returns None on error

Example

let veeks = veeks_millis::VeekDate::from_str("2021c 09v 3d").unwrap();
assert_eq!(veeks.to_string(),"2021c 09v 3d");

pub fn year(&self) -> u32

getter year u32

Example

let veeks = veeks_millis::VeekDate::from_str("2021c 09v 3d").unwrap();
assert_eq!(veeks.year(),2021);

pub fn veek(&self) -> u32

getter veek u32

Example

let veeks = veeks_millis::VeekDate::from_str("2021c 09v 3d").unwrap();
assert_eq!(veeks.veek(),9);

pub fn day(&self) -> u32

getter day u32

Example

let veeks = veeks_millis::VeekDate::from_str("2021c 09v 3d").unwrap();
assert_eq!(veeks.day(),3);

pub fn to_naive_date(&self) -> Option<NaiveDate>

convert veek_date to chrono NaiveDate returns None on error

Example

let vd = veeks_millis::VeekDate::from_ymd(2021,02,28).unwrap();
let nd = vd.to_naive_date().unwrap();
assert_eq!(nd, chrono::NaiveDate::from_ymd(2021,02,28));